fix: move in closed loop
Calculation of desired absolute position when motor is in closed loop may
be wrong. This could happen when POS AXIS and POS [ENC] e.g. POS ABSENC
registers have diverged. This happens because the desired absolute
position uses both registers, first the POS [ENC] to calculate
the necessary increment and second the POS AXIS to calculate
the absolute position.  And the MOVE request in closed loop expects
an absolute position based on encoder (POS [ENC]).
Fix it by avoiding this "two-steps calculation" in case of closed loop.